+++ packages/orq-rc/src/models/components/conversationresponse.ts
@@ -131,7 +131,7 @@ export const ConversationResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kff8m4a87b28xnf42dpves6g"),
+  _id: z.string().default("conv_01kfg037949f418k8a99tx43tw"),
   kind: ConversationResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -162,7 +162,7 @@ export const ConversationResponse$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ConversationResponse
 > = z.object({
-  id: z.string().default("conv_01kff8m4a87b28xnf42dpves6g"),
+  id: z.string().default("conv_01kfg037949f418k8a99tx43tw"),
   kind: ConversationResponseKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),