+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -270,7 +270,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kff8m4abwqsfx53c76twafww"),
+  _id: z.string().default("conv_01kfg037957qts609ny87x1bp8"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -302,7 +302,7 @@ export const CreateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kff8m4abwqsfx53c76twafww"),
+  id: z.string().default("conv_01kfg037957qts609ny87x1bp8"),
   kind: CreateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),