+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -316,7 +316,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kff8m4amwwewf3bgmkm8wbd9"),
+  _id: z.string().default("conv_01kfg0379cq05pra7z81qp3hcf"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -348,7 +348,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kff8m4amwwewf3bgmkm8wbd9"),
+  id: z.string().default("conv_01kfg0379cq05pra7z81qp3hcf"),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),