+++ packages/orq-rc/RELEASES.md
@@ -6069,3 +6069,13 @@ Based on:
 - [typescript v4.2.0-rc.53] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.53] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.53 - packages/orq-rc
+
+## 2026-01-21 10:07:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.691.0 (2.797.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.54] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.54] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.54 - packages/orq-rc